Create main.py
when you first run the script a webpage will be opend fo you to Authorize the account. when taken to the "http://example.com" site, copy the entire URL and paste it into the terminal where it asks for the URL. 
This will get the neccisery tokens and user_ids.

Usually you only need to do this once. After that, enter the date in the format shown and press enter. It will now create a playlist and add all the songs to that playlist using the URI